P008 fix: switch to double not int

+++ Projects/008-complex-number-class/complex.h
@@ -3,37 +3,37 @@
 
 class Complex {
     private:
-        int real;
-        int img;
+        double real;
+        double img;
     public:
-        Complex(int real = 0, int img = 0);
+        Complex(double real = 0, double img = 0);
 
-        void setReal(int real);
-        int getReal() const;
-        void setImg(int img);
-        int getImg() const;
+        void setReal(double real);
+        double getReal() const;
+        void setImg(double img);
+        double getImg() const;
         void displayComplex();
 
         Complex operator+(const Complex &other) const;
         Complex operator-(const Complex &other) const;
         Complex operator*(const Complex &other) const;
         Complex operator/(const Complex &other) const;
-        Complex operator+(int n) const;
-        Complex operator-(int n) const;
-        Complex operator*(int n) const;
-        Complex operator/(int n) const;
+        Complex operator+(double n) const;
+        Complex operator-(double n) const;
+        Complex operator*(double n) const;
+        Complex operator/(double n) const;
         void operator=(const Complex &other);
         void operator+=(const Complex &other);
         void operator-=(const Complex &other);
         void operator*=(const Complex &other);
         void operator/=(const Complex &other);
-        void operator+=(int n);
-        void operator-=(int n);
-        void operator*=(int n);
-        void operator/=(int n);
+        void operator+=(double n);
+        void operator-=(double n);
+        void operator*=(double n);
+        void operator/=(double n);
         bool operator==(const Complex &other) const;
         bool operator!=(const Complex &other) const;
-        int magnitude() const;
+        double magnitude() const;
         Complex conjugate() const;
 
         // reverse
